requirements.txt
maya.cmds
mayapy